05.10.2003 Bugfixes for subsetted Fonts and Drag&Drop of Objects
git-svn-id: svn://scribus.net/trunk/Scribus@75 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/mspinbox.cpp
@@ -67,21 +67,21 @@ bool MSpinBox::eventFilter( QObject* ob, QEvent* ev )
 
 QString MSpinBox::mapValueToText(int value)
 {
-	int dez;
+	double dez;
 	if (Width == 1)
-		dez = 10;
+		dez = 10.0;
 	else
-		dez = 100;
+		dez = 100.0;
 	return QString::number(static_cast<double>(value) / dez, 'f', Width);
 }
 
 int MSpinBox::mapTextToValue(bool *)
 {
-	int dez;
+	float dez;
 	if (Width == 1)
-		dez = 10;
+		dez = 10.0;
 	else
-		dez = 100;
+		dez = 100.0;
 	return int(qRound(text().toFloat()*dez));
 }
